flacparse: skip padding metadata blocks
Additionally, add a skip-padding-metadata property to
opt-out of this default behavior.

Padding metadata blocks are in-place editing
placeholders left by an enconder to facilitate
addition of extra metadata without having to re-master
the entire file. While this makes them useless during
playback, an user might want to preserve them during
stream copying. This is why the new behavior - skipping
them - is introduced as an oup-out default.

https://bugzilla.gnome.org/show_bug.cgi?id=759778